% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBERlibR.R
\name{gui}
\alias{gui}
\title{Graphic User Interface of DBERlibR}
\usage{
gui()
}
\description{
This function is just load a graphical user interface of DBERlibR.
Run gui() in the console and the GUI will pop up.
}
