% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBERlibR.R
\name{independentsamples}
\alias{independentsamples}
\title{Independent Samples Data Analysis}
\usage{
independentsamples()
}
\description{
This function automatically cleans the data sets (e.g., converting missing values to "0), binds treatment-control group Data Sets, runs the Independent Samples T-test (parametric) and Mann–Whitney U test (nonparametric), and then exports outputs to help you examine the difference between the groups. R scripts and their outputs are as follows (just pay attention to the outputs since the codes are automatically run back-end by the function). Find the output file named "independentsamples_output.txt" instantly generated by this function in the working directory.
}
