% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBERlibR.R
\name{itemanalysis}
\alias{itemanalysis}
\title{Item Analysis}
\usage{
itemanalysis(csv_data)
}
\arguments{
\item{csv_data}{This function requires a csv data file name: "data_treat_pre.csv", "data_treat_post.csv", "data_ctrl_pre.csv", or "data_ctrl_post.csv".}
}
\description{
This function automatically reads and cleans the data (e.g., converting missing values to "0"), calculates difficulty and discriminant scores, and exports the results to an Excel file, which is saved in the working directory. The Excel file name follows the data file name used for item analysis; for example, if you've input "data_treat_pre.csv" in the function, then the output Excel file name contains "treat_pre." The function also generates plots in the jpeg file format to visualize the results; the jpeg file name follows the data file name as well so that you can easily find the results of the item analysis you've just run. Find the output files of which name begins with "itemanalysis" instantly generated by this function in the working directory.
}
