% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBERlibR.R
\name{multi_to_binary}
\alias{multi_to_binary}
\title{conversion of multiple-choice answers to a binary format of 1 and 0}
\usage{
multi_to_binary(csv_list)
}
\description{
This function automatically converts multiple-choice answers to a binary format of "1" for correct answers and "0" for incorrect answers. Users need to create a separate CSV file with answer keys for all questions (including true/false questions) in the data file(s) and save the file as "answer_keys.csv." The variable name for answer keys should be "key," as instructed in the tutorial.
}
