% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBERlibR.R
\name{one_way_ancova}
\alias{one_way_ancova}
\title{One-way ANCOVA}
\usage{
one_way_ancova()
}
\description{
This function automatically merges pre-post datasets, binds treatment-control datasets, and runs scripts to check assumptions of one-way ANCOVA, runs the main One-way ANCOVA, and then exports all outputs for you all at once. Please make sure to name data files accurately (i.e., “data_treat_pre.csv”, “data_treat_post.csv”, “data_ctrl_pre.csv”," data_ctrl_post.csv") and have them saved in the working directory. Find the output files "onewayancova_outputs.txt" and "onewayancova_linearity_check_scatter_plot.png" instantly generated by this function in the working directory.
}
