% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBERlibR.R
\name{one_way_repeated_anova}
\alias{one_way_repeated_anova}
\title{One-way Repeated Measures ANOVA}
\usage{
one_way_repeated_anova()
}
\description{
This function automatically merges pre, post, and post2 datasets, runs the one-way repeated measures ANOVA with assumptions check, and then exports outputs for you all at once. Please make sure to name data files accurately (i.e., “data_treat_pre.csv”, “data_treat_post.csv”, and “data_treat_post2.csv”) and have them saved in the working directory. Find the output files "onewayrepeatedanova_outputs.txt" instantly generated by this function in the working directory.
}
