% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBERlibR.R
\name{onewayancova}
\alias{onewayancova}
\title{One-way ANCOVA}
\usage{
onewayancova()
}
\description{
This function automatically merges pre-post data sets, binds treatment-control data sets, and runs scripts to check assumptions of one-way ANCOVA, runs the main One-way ANCOVA, and then exports all outputs for you all at once. Please make sure to name data files accurately (i.e., “data_treat_pre.csv”, “data_treat_post.csv”, “data_ctrl_pre.csv”," data_ctrl_post.csv") and have them saved in the working directory. Find the output files "onewayancova_outputs.txt" and "onewayancova_linearity_check_scatter_plot.jpeg" instantly generated by this function in the working directory.
}
