% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBERlibR.R
\name{onewayrepeatedanova}
\alias{onewayrepeatedanova}
\title{One-way Repeated Measures ANOVA}
\usage{
onewayrepeatedanova()
}
\description{
This function automatically merges pre, post, and post2 data sets, runs the one-way repeated measures ANOVA with assumptions check, and then exports outputs for you all at once. Please make sure to name data files accurately (i.e., “data_treat_pre.csv”, “data_treat_post.csv”, and “data_treat_post2.csv”) and have them saved in the working directory. Find the output files "onewayrepeatedanova_outputs.txt" instantly generated by this function in the working directory.
}
