% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBERlibR.R
\name{paired_samples}
\alias{paired_samples}
\title{Paired Samples Data Analysis}
\usage{
paired_samples()
}
\description{
This function automatically cleans the datasets (e.g., converting missing values to "0), merges pre-post datasets, runs the (parametric) Paired Samples T-test and (nonparametric) Wilcoxon Signed-Rank test, and then exports outputs to help you examine the difference between pre-post scores. Find the output files named as "pairedsamples_output.txt" and "pairedsamples_invidual_items.png" instantly created by this function in the working directory.
}
