% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBERlibR.R
\name{pairedsamples}
\alias{pairedsamples}
\title{Paired Samples Data Analysis}
\usage{
pairedsamples()
}
\description{
This function automatically cleans the data sets (e.g., converting missing values to "0), merges pre-post Data Sets, runs the (parametric) Paired Samples T-test and (nonparametric) Wilcoxon Signed-Rank test, and then exports outputs to help you examine the difference between pre-post scores. Find the output files named as "pairedsamples_output.txt" and "pairedsamples_invidual_items.jpeg" instantly created by this function in the working directory.
}
