% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBERlibR.R
\name{two_way_repeated_anova}
\alias{two_way_repeated_anova}
\title{two-way Repeated Measures ANOVA}
\usage{
two_way_repeated_anova()
}
\description{
This function automatically merges pre, post, and post2 datasets, binds intervention and control group datasets, runs the two-way repeated measures ANOVA with assumptions check, and then displays/exports outputs all at once. Please make sure to name data files accurately (i.e., “data_treat_pre.csv”, “data_treat_post.csv”, and “data_treat_post2.csv”, “data_ctrl_pre.csv”, “data_ctrl_post.csv”, and “data_ctrl_post2.csv”) and have them saved in the working directory. Find the output files "twowayrepeatedanova_outputs.txt" instantly generated by this function in the working directory.
}
